refactoring logic into services, initial ingest service implementation (#1)
